Add files via upload
basic shit, utilizes python 3. If you want to run it, copy and paste file into a text file with a .py extension. Once this is done go to terminal, go to the file location, and then type in "python3 (filename).py" and you should get the basic mining message. Work in progress but it works.